+++ Eform1/.vs/Eform1/xs/UserPrefs.xml
@@ -1,10 +1,29 @@
 ﻿<Properties StartupConfiguration="{D73BDF18-3D4A-4090-BC11-251C4302D678}|Default">
-  <MonoDevelop.Ide.Workbench>
+  <MonoDevelop.Ide.Workbench ActiveDocument="models/MySQLFormRepository.cs">
+    <Files>
+      <File FileName="Data/FormDbContext.cs" Line="27" Column="1" />
+      <File FileName="models/table_1.cs" Line="31" Column="1" />
+      <File FileName="models/table_2.cs" Line="38" Column="1" />
+      <File FileName="models/table_3.cs" Line="40" Column="1" />
+      <File FileName="Startup.cs" Line="21" Column="53" />
+      <File FileName="appsettings.json" Line="14" Column="1" />
+      <File FileName="models/IFormRepository.cs" Line="1" Column="1" />
+      <File FileName="models/Formdumy.cs" Line="1" Column="1" />
+      <File FileName="controllers/HomeController.cs" Line="1" Column="1" />
+      <File FileName="models/MySQLFormRepository.cs" Line="1" Column="1" />
+    </Files>
     <Pads>
       <Pad Id="ProjectPad">
         <State name="__root__">
-          <Node name="Eform1" expanded="True" selected="True">
-            <Node name="Eform1" expanded="True" />
+          <Node name="Eform1" expanded="True">
+            <Node name="Eform1" expanded="True">
+              <Node name="controllers" expanded="True" />
+              <Node name="Data" expanded="True" />
+              <Node name="models" expanded="True">
+                <Node name="MySQLFormRepository.cs" selected="True" />
+              </Node>
+              <Node name="Views" expanded="True" />
+            </Node>
           </Node>
         </State>
       </Pad>
