+++ Eform1/bin/Debug/netcoreapp3.1/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "FormDbConnection": "server=localhost;port=3306;database=eform;user=root;password=jayaram2000"
+    "FormDbConnection": "server=localhost;port=3306;database=eform;user=root;password=990126@Ub"
   }
 }
