+++ Eform1/models/table_2.cs
@@ -9,17 +9,28 @@ namespace Eform1.models
 {
     public class table_2
     {
-        [Key]
-       // public int DUM { get; set; }
-        
+
+        // public int DUM { get; set; }
+
         //[ForeignKey("table_1")]
+
+        [ForeignKey("table_1")]
         public int UID_F1 { get; set; }
-        
+
+        [Key]
+        [Required]
         public int UID_Q { get; set; }
+
         public string Question_Type { get; set; }
+
         public string Question { get; set; }
 
-       // public table_1 table_1 { get; set; }
+        #region Navigation Properties
+        public table_1 table_1 { get; set; }
+        public ICollection<table_3> table_3 { get; set; }
+        #endregion
+
+        // public table_1 table_1 { get; set; }
         //public table_3 table_3 { get; set; }  [ForeignKey("table_3")]
     }
 }
