+++ Eform1/models/table_3.cs
@@ -10,21 +10,29 @@ namespace Eform1.models
     public class table_3
     {
         //[Key]
-       // [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        // [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         //public Guid idtable_3s { get; set; }
-       // [Key]
-      //  public int idtable_3s { get; set; }
+        // [Key]
+        //  public int idtable_3s { get; set; }
 
         //[ForeignKey("table_2")]
-       [Key]
+
+        // public int DUM1 { get; set; }
+
         public int F2 { get; set; }
 
-       
-        // public int DUM1 { get; set; }
-        
+        [ForeignKey("table_2")]
         public int UID_Q { get; set; }
+
+        [Key]
+        [Required]
         public int ID_MCQ { get; set; }
+
         public string Options { get; set; }
+
+        #region Navigation Properties
+        public table_2 table_2 { get; set; }
+        #endregion
 
         //public table_4 table_4 { get; set; } [ForeignKey("table_4")]
     }
